CCM-8023: style fixes, new oauth2 callback page
