+++ infrastructure/terraform/components/app/cognito_user_pool_client.tf
@@ -7,8 +7,8 @@ resource "aws_cognito_user_pool_client" "main" {
   callback_urls = flatten([
     var.cognito_user_pool_additional_callback_urls,
     [
-      "https://${var.environment}.${local.acct.dns_zone["name"]}/auth",
-      "https://${aws_amplify_app.main.default_domain}/auth"
+      "https://${var.environment}.${local.acct.dns_zone["name"]}/auth/oauth2",
+      "https://${aws_amplify_app.main.default_domain}/auth/oauth2"
     ],
     var.cognito_user_pool_use_environment_specific_gateway_callback_url ? [
       "https://${var.environment}.${var.cognito_user_pool_environment_specific_gateway_callback_url_suffix}"
@@ -16,13 +16,13 @@ resource "aws_cognito_user_pool_client" "main" {
   ])
 
   logout_urls = flatten([
-    var.cognito_user_pool_additional_callback_urls,
+    var.cognito_user_pool_additional_logout_urls,
     [
       "https://${var.environment}.${local.acct.dns_zone["name"]}/auth",
       "https://${aws_amplify_app.main.default_domain}/auth"
     ],
     var.cognito_user_pool_use_environment_specific_gateway_callback_url ? [
-      "https://${var.environment}.${var.cognito_user_pool_environment_specific_gateway_callback_url_suffix}"
+      "https://${var.environment}.${var.cognito_user_pool_environment_specific_gateway_logout_url_suffix}"
     ] : []
   ])
