+++ src/app/page.tsx
@@ -1,14 +1,13 @@
 'use client';
 
-import React, { Suspense, useEffect, useState } from 'react';
-import path from 'path';
-import { redirect, RedirectType, useSearchParams } from 'next/navigation';
-import { Hub } from 'aws-amplify/utils';
-import { Authenticator, useAuthenticator } from '@aws-amplify/ui-react';
-import { federatedSignIn, State } from '@/src/utils/federated-sign-in';
+import React, { Suspense } from 'react';
+import { useSearchParams } from 'next/navigation';
+import { Authenticator } from '@aws-amplify/ui-react';
+import { federatedSignIn } from '@/src/utils/federated-sign-in';
 import { CIS2SignInButton } from '@/src/components/CIS2SignInButton/CIS2SignInButton';
 import content from '@/src/content/content';
 import styles from './page.module.scss';
+import { Redirect } from '../components/molecules/Redirect/Redirect';
 
 function SignInPage() {
   const {
@@ -16,33 +15,7 @@ function SignInPage() {
   } = content;
   const searchParams = useSearchParams();
 
-  const [customState, setCustomState] = useState<State>();
-
   const redirectPath = searchParams.get('redirect');
-
-  const { authStatus } = useAuthenticator((ctx) => [ctx.authStatus]);
-
-  useEffect(() => {
-    return Hub.listen('auth', ({ payload }) => {
-      if (payload.event === 'customOAuthState') {
-        setCustomState(JSON.parse(payload.data));
-      }
-    });
-  }, []);
-
-  useEffect(() => {
-    let to = '';
-
-    if (customState) {
-      to = customState.redirectPath || '/home';
-    } else if (authStatus === 'authenticated' && redirectPath) {
-      to = redirectPath;
-    }
-
-    if (to) {
-      redirect(path.normalize(`/redirect/${to}`), RedirectType.replace);
-    }
-  }, [authStatus, customState, redirectPath]);
 
   return (
     <div className='nhsuk-grid-row'>
@@ -65,7 +38,9 @@ function SignInPage() {
               <h2 className='nhsuk-heading-m'>
                 {content.components.cognitoSignInComponent.heading}
               </h2>
-              <Authenticator variation='default' hideSignUp />
+              <Authenticator variation='default' hideSignUp>
+                <Redirect />
+              </Authenticator>
             </div>
           </>
         )}
